fix(parser): change port type from short to int to avoid overflow

+++ includes/http-request-parser/configuration-file/ListenData.hpp
@@ -6,7 +6,7 @@
 
 struct ListenData
 {
-	short		port;
+	int			port;
 	std::string	ip_address;
 
 	bool operator==(const ListenData& other) const;
